Create OfferClosed Handler. Move functionality away from newly created WebhookService and delete it
