+++ Helper/PaymentMethods.php
@@ -12,9 +12,20 @@
 namespace Adyen\Payment\Helper;
 
 use Adyen\AdyenException;
+use Adyen\Payment\Logger\AdyenLogger;
+use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\App\Helper\AbstractHelper;
+use Magento\Framework\App\Helper\Context;
+use Magento\Framework\App\RequestInterface;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Framework\Locale\ResolverInterface;
+use Magento\Framework\View\Asset\Repository;
+use Magento\Framework\View\Asset\Source;
+use Magento\Framework\View\Design\Theme\ThemeProviderInterface;
+use Magento\Framework\View\DesignInterface;
+use Magento\Quote\Api\CartRepositoryInterface;
+use Magento\Sales\Model\Order;
 
 /**
  * @SuppressWarnings(PHPMD.LongVariable)
@@ -34,27 +45,27 @@ class PaymentMethods extends AbstractHelper
     ];
 
     /**
-     * @var \Magento\Quote\Api\CartRepositoryInterface
+     * @var CartRepositoryInterface
      */
     protected $quoteRepository;
 
     /**
-     * @var \Magento\Framework\App\Config\ScopeConfigInterface $config
+     * @var ScopeConfigInterface $config
      */
     protected $config;
 
     /**
-     * @var \Adyen\Payment\Helper\Data
+     * @var Data
      */
     protected $adyenHelper;
 
     /**
-     * @var \Magento\Framework\Locale\ResolverInterface
+     * @var ResolverInterface
      */
     protected $localeResolver;
 
     /**
-     * @var \Adyen\Payment\Logger\AdyenLogger
+     * @var AdyenLogger
      */
     protected $adyenLogger;
 
@@ -69,17 +80,17 @@ class PaymentMethods extends AbstractHelper
     protected $request;
 
     /**
-     * @var \Magento\Framework\View\Asset\Source
+     * @var Source
      */
     protected $assetSource;
 
     /**
-     * @var \Magento\Framework\View\DesignInterface
+     * @var DesignInterface
      */
     protected $design;
 
     /**
-     * @var \Magento\Framework\View\Design\Theme\ThemeProviderInterface
+     * @var ThemeProviderInterface
      */
     protected $themeProvider;
 
@@ -93,33 +104,42 @@ class PaymentMethods extends AbstractHelper
      */
     private $chargedCurrency;
 
+    /** @var Config */
+    private $configHelper;
+
     /**
      * PaymentMethods constructor.
      *
-     * @param \Magento\Quote\Api\CartRepositoryInterface $quoteRepository
-     * @param \Magento\Framework\App\Config\ScopeConfigInterface $config
+     * @param Context $context
+     * @param CartRepositoryInterface $quoteRepository
+     * @param ScopeConfigInterface $config
      * @param Data $adyenHelper
-     * @param \Magento\Framework\Locale\ResolverInterface $localeResolver
-     * @param \Adyen\Payment\Logger\AdyenLogger $adyenLogger
-     * @param \Magento\Framework\View\Asset\Repository $assetRepo
-     * @param \Magento\Framework\App\RequestInterface $request
-     * @param \Magento\Framework\View\Asset\Source $assetSource
-     * @param \Magento\Framework\View\DesignInterface $design
-     * @param \Magento\Framework\View\Design\Theme\ThemeProviderInterface $themeProvider
+     * @param ResolverInterface $localeResolver
+     * @param AdyenLogger $adyenLogger
+     * @param Repository $assetRepo
+     * @param RequestInterface $request
+     * @param Source $assetSource
+     * @param DesignInterface $design
+     * @param ThemeProviderInterface $themeProvider
+     * @param ChargedCurrency $chargedCurrency
+     * @param Config $configHelper
      */
     public function __construct(
-        \Magento\Quote\Api\CartRepositoryInterface $quoteRepository,
-        \Magento\Framework\App\Config\ScopeConfigInterface $config,
-        \Adyen\Payment\Helper\Data $adyenHelper,
-        \Magento\Framework\Locale\ResolverInterface $localeResolver,
-        \Adyen\Payment\Logger\AdyenLogger $adyenLogger,
-        \Magento\Framework\View\Asset\Repository $assetRepo,
-        \Magento\Framework\App\RequestInterface $request,
-        \Magento\Framework\View\Asset\Source $assetSource,
-        \Magento\Framework\View\DesignInterface $design,
-        \Magento\Framework\View\Design\Theme\ThemeProviderInterface $themeProvider,
-        ChargedCurrency $chargedCurrency
+        Context $context,
+        CartRepositoryInterface $quoteRepository,
+        ScopeConfigInterface $config,
+        Data $adyenHelper,
+        ResolverInterface $localeResolver,
+        AdyenLogger $adyenLogger,
+        Repository $assetRepo,
+        RequestInterface $request,
+        Source $assetSource,
+        DesignInterface $design,
+        ThemeProviderInterface $themeProvider,
+        ChargedCurrency $chargedCurrency,
+        Config $configHelper
     ) {
+        parent::__construct($context);
         $this->quoteRepository = $quoteRepository;
         $this->config = $config;
         $this->adyenHelper = $adyenHelper;
@@ -131,6 +151,7 @@ class PaymentMethods extends AbstractHelper
         $this->design = $design;
         $this->themeProvider = $themeProvider;
         $this->chargedCurrency = $chargedCurrency;
+        $this->configHelper = $configHelper;
     }
 
     /**
@@ -260,7 +281,7 @@ class PaymentMethods extends AbstractHelper
         return "";
     }
 
-     /**
+    /**
      * @param $requestParams
      * @param $store
      * @return array
@@ -514,5 +535,189 @@ class PaymentMethods extends AbstractHelper
         ];
 
         return in_array($paymentMethod, $paymentMethodRecurring);
+    }
+
+    /**
+     * Check if order should be automatically captured
+     *
+     * @param Order $order
+     * @param string $notificationPaymentMethod
+     * @return bool
+     */
+    public function isAutoCapture(Order $order, string $notificationPaymentMethod): bool
+    {
+        // validate if payment methods allows manual capture
+        if ($this->manualCaptureAllowed($notificationPaymentMethod)) {
+            $captureMode = trim(
+                $this->configHelper->getConfigData(
+                    'capture_mode',
+                    'adyen_abstract',
+                    $order->getStoreId()
+                )
+            );
+            $sepaFlow = trim(
+                $this->configHelper->getConfigData(
+                    'sepa_flow',
+                    'adyen_abstract',
+                    $order->getStoreId()
+                )
+            );
+            $paymentCode = $order->getPayment()->getMethod();
+            $captureModeOpenInvoice = $this->configHelper->getConfigData(
+                'auto_capture_openinvoice',
+                'adyen_abstract',
+                $order->getStoreId()
+            );
+            $manualCapturePayPal = trim(
+                $this->configHelper->getConfigData(
+                    'paypal_capture_mode',
+                    'adyen_abstract',
+                    $order->getStoreId()
+                )
+            );
+
+            /*
+             * if you are using authcap the payment method is manual.
+             * There will be a capture send to indicate if payment is successful
+             */
+            if ($notificationPaymentMethod == "sepadirectdebit") {
+                if ($sepaFlow == "authcap") {
+                    $this->adyenLogger->addAdyenNotificationCronjob(
+                        'Manual Capture is applied for sepa because it is in authcap flow'
+                    );
+                    return false;
+                } else {
+                    // payment method ideal, cash adyen_boleto has direct capture
+                    $this->adyenLogger->addAdyenNotificationCronjob(
+                        'This payment method does not allow manual capture.(2) paymentCode:' .
+                        $paymentCode . ' paymentMethod:' . $notificationPaymentMethod . ' sepaFLow:' . $sepaFlow
+                    );
+                    return true;
+                }
+            }
+
+            if ($paymentCode == "adyen_pos_cloud") {
+                $captureModePos = $this->adyenHelper->getAdyenPosCloudConfigData(
+                    'capture_mode_pos',
+                    $order->getStoreId()
+                );
+                if (strcmp($captureModePos, 'auto') === 0) {
+                    $this->adyenLogger->addAdyenNotificationCronjob(
+                        'This payment method is POS Cloud and configured to be working as auto capture '
+                    );
+                    return true;
+                } elseif (strcmp($captureModePos, 'manual') === 0) {
+                    $this->adyenLogger->addAdyenNotificationCronjob(
+                        'This payment method is POS Cloud and configured to be working as manual capture '
+                    );
+                    return false;
+                }
+            }
+
+            // if auto capture mode for openinvoice is turned on then use auto capture
+            if ($captureModeOpenInvoice &&
+                $this->adyenHelper->isPaymentMethodOpenInvoiceMethodValidForAutoCapture($notificationPaymentMethod)
+            ) {
+                $this->adyenLogger->addAdyenNotificationCronjob(
+                    'This payment method is configured to be working as auto capture '
+                );
+                return true;
+            }
+
+            // if PayPal capture modues is different from the default use this one
+            if (strcmp($notificationPaymentMethod, 'paypal') === 0) {
+                if ($manualCapturePayPal) {
+                    $this->adyenLogger->addAdyenNotificationCronjob(
+                        'This payment method is paypal and configured to work as manual capture'
+                    );
+                    return false;
+                } else {
+                    $this->adyenLogger->addAdyenNotificationCronjob(
+                        'This payment method is paypal and configured to work as auto capture'
+                    );
+                    return true;
+                }
+            }
+            if (strcmp($captureMode, 'manual') === 0) {
+                $this->adyenLogger->addAdyenNotificationCronjob(
+                    'Capture mode for this payment is set to manual'
+                );
+                return false;
+            }
+
+            /*
+             * online capture after delivery, use Magento backend to online invoice
+             * (if the option auto capture mode for openinvoice is not set)
+             */
+            if ($this->adyenHelper->isPaymentMethodOpenInvoiceMethod($notificationPaymentMethod)) {
+                $this->adyenLogger->addAdyenNotificationCronjob(
+                    'Capture mode for klarna is by default set to manual'
+                );
+                return false;
+            }
+
+            $this->adyenLogger->addAdyenNotificationCronjob('Capture mode is set to auto capture');
+            return true;
+        } else {
+            // does not allow manual capture so is always immediate capture
+            $this->adyenLogger->addAdyenNotificationCronjob(
+                sprintf('Payment method %s, does not allow manual capture', $notificationPaymentMethod)
+            );
+
+            return true;
+        }
+    }
+
+    /**
+     * Validate if this payment methods allows manual capture
+     * This is a default can be forced differently to overrule on acquirer level
+     *
+     * @param string $notificationPaymentMethod
+     * @return bool
+     */
+    private function manualCaptureAllowed(string $notificationPaymentMethod): bool
+    {
+        $manualCaptureAllowed = false;
+        // For all openinvoice methods manual capture is the default
+        if ($this->adyenHelper->isPaymentMethodOpenInvoiceMethod($notificationPaymentMethod)) {
+            return true;
+        }
+
+        switch ($notificationPaymentMethod) {
+            case 'cup':
+            case 'cartebancaire':
+            case 'visa':
+            case 'visadankort':
+            case 'mc':
+            case 'uatp':
+            case 'amex':
+            case 'maestro':
+            case 'maestrouk':
+            case 'diners':
+            case 'discover':
+            case 'jcb':
+            case 'laser':
+            case 'paypal':
+            case 'sepadirectdebit':
+            case 'dankort':
+            case 'elo':
+            case 'hipercard':
+            case 'mc_applepay':
+            case 'visa_applepay':
+            case 'amex_applepay':
+            case 'discover_applepay':
+            case 'maestro_applepay':
+            case 'paywithgoogle':
+            case 'svs':
+            case 'givex':
+            case 'valuelink':
+            case 'twint':
+                $manualCaptureAllowed = true;
+                break;
+            default:
+                break;
+        }
+
+        return $manualCaptureAllowed;
     }
 }
